requirements.txt
# AI Usage Analytics Dashboard Dependencies
# Core Python libraries
google-cloud-bigquery>=3.0.0
google-cloud-secret-manager>=2.0.0
google-cloud-logging>=3.0.0
google-cloud-monitoring>=2.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
pandas>=2.0.0
requests>=2.28.0

# Testing and development
pytest>=7.4.0
pytest-mock>=3.10.0
black>=23.0.0
flake8>=6.0.0

# Web framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.1